+++ TWLight/resources/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "TWLight.resources.app.Config"
