+++ TWLight/resources/models.py
@@ -14,8 +14,6 @@ from django.core.urlresolvers import reverse_lazy, reverse
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django_countries.fields import CountryField
-from django.dispatch import receiver
-from django.db.models.signals import post_delete
 
 
 RESOURCE_LANGUAGES = copy.copy(settings.INTERSECTIONAL_LANGUAGES)
@@ -672,18 +670,6 @@ class Stream(models.Model):
         elif self.target_url:
             access_url = self.target_url
         return access_url
-
-
-@receiver(post_delete, sender=Stream)
-def set_partner_specific_stream_false_if_no_streams(sender, instance, **kwargs):
-    """
-    Set parent partner's specific_stream to False when the last stream is deleted.
-    """
-
-    partner = instance.partner
-    if partner.specific_stream and Stream.objects.filter(partner=partner).count() == 0:
-        partner.specific_stream = False
-        partner.save()
 
 
 class Contact(models.Model):
