+++ TWLight/users/signals.py
@@ -1,8 +1,9 @@
 from datetime import timedelta
+from django.conf import settings
 from django.contrib.auth.models import User
 from django.dispatch import receiver, Signal
 from django.db.models.signals import post_save, post_delete
-from TWLight.users.models import Authorization
+from TWLight.users.models import Authorization, UserProfile
 from TWLight.resources.models import Partner, Stream
 
 
@@ -16,6 +17,13 @@ class Notice(object):
             "partner_link",
         ]
     )
+
+
+@receiver(post_save, sender=settings.AUTH_USER_MODEL)
+def create_user_profile(sender, instance, created, **kwargs):
+    """Create user profiles automatically when users are created."""
+    if created:
+        UserProfile.objects.create(user=instance)
 
 
 @receiver(post_save, sender=Partner)
@@ -66,6 +74,7 @@ def delete_all_but_latest_partner_authorizations(sender, instance, **kwargs):
     partner = instance.partner
     authorizations = Authorization.objects.filter(partner=partner)
     users = User.objects.filter(authorization__in=authorizations)
+
     for user in users:
         partner_authorizations = Authorization.objects.filter(
             user=user, partner=partner, stream__isnull=True
